# System.Linq ToDictionary performance improvement