--- a/src/System.Linq/src/System/Linq/Enumerable.cs
+++ b/src/System.Linq/src/System/Linq/Enumerable.cs
@@ -1032,14 +1032,39 @@ namespace System.Linq
             if (keySelector == null) throw Error.ArgumentNull("keySelector");
 
             int capacity = 0;
+            ICollection<TSource> collection = source as ICollection<TSource>;
+            if (collection != null)
+            {
+                capacity = collection.Count;
+                if (capacity == 0)
+                    return new Dictionary<TKey, TSource>(comparer);
 
-            if (source is TSource[]) capacity = (source as TSource[]).Length; // Cast to Array is much faster than cast to ICollection
-            else if (source is ICollection<TSource>) capacity = (source as ICollection<TSource>).Count;
+                TSource[] array = collection as TSource[];
+                if (array != null)
+                    return ToDictionary(array, keySelector, comparer);
+                List<TSource> list = collection as List<TSource>;
+                if (list != null)
+                    return ToDictionary(list, keySelector, comparer);
+            }
 
             Dictionary<TKey, TSource> d = new Dictionary<TKey, TSource>(capacity, comparer);
             foreach (TSource element in source) d.Add(keySelector(element), element);
             return d;
         }
+
+        private static Dictionary<TKey, TSource> ToDictionary<TSource, TKey>(TSource[] source, Func<TSource, TKey> keySelector, IEqualityComparer<TKey> comparer)
+        {
+            Dictionary<TKey, TSource> d = new Dictionary<TKey, TSource>(source.Length, comparer);
+            for (int i = 0; i < source.Length; i++) d.Add(keySelector(source[i]), source[i]);
+            return d;
+        }
+        private static Dictionary<TKey, TSource> ToDictionary<TSource, TKey>(List<TSource> source, Func<TSource, TKey> keySelector, IEqualityComparer<TKey> comparer)
+        {
+            Dictionary<TKey, TSource> d = new Dictionary<TKey, TSource>(source.Count, comparer);
+            foreach (TSource element in source) d.Add(keySelector(element), element);
+            return d;
+        }
+
 
         public static Dictionary<TKey, TElement> ToDictionary<TSource, TKey, TElement>(this IEnumerable<TSource> source, Func<TSource, TKey> keySelector, Func<TSource, TElement> elementSelector)
         {
@@ -1053,14 +1078,39 @@ namespace System.Linq
             if (elementSelector == null) throw Error.ArgumentNull("elementSelector");
 
             int capacity = 0;
+            ICollection<TSource> collection = source as ICollection<TSource>;
+            if (collection != null)
+            {
+                capacity = collection.Count;
+                if (capacity == 0)
+                    return new Dictionary<TKey, TElement>(comparer);
 
-            if (source is TSource[]) capacity = (source as TSource[]).Length; // Cast to Array is much faster than cast to ICollection
-            else if (source is ICollection<TSource>) capacity = (source as ICollection<TSource>).Count;
+                TSource[] array = collection as TSource[];
+                if (array != null)
+                    return ToDictionary(array, keySelector, elementSelector, comparer);
+                List<TSource> list = collection as List<TSource>;
+                if (list != null)
+                    return ToDictionary(list, keySelector, elementSelector, comparer);
+            }
 
             Dictionary<TKey, TElement> d = new Dictionary<TKey, TElement>(capacity, comparer);
             foreach (TSource element in source) d.Add(keySelector(element), elementSelector(element));
             return d;
         }
+
+        private static Dictionary<TKey, TElement> ToDictionary<TSource, TKey, TElement>(TSource[] source, Func<TSource, TKey> keySelector, Func<TSource, TElement> elementSelector, IEqualityComparer<TKey> comparer)
+        {
+            Dictionary<TKey, TElement> d = new Dictionary<TKey, TElement>(source.Length, comparer);
+            for (int i = 0; i < source.Length; i++) d.Add(keySelector(source[i]), elementSelector(source[i]));
+            return d;
+        }
+        private static Dictionary<TKey, TElement> ToDictionary<TSource, TKey, TElement>(List<TSource> source, Func<TSource, TKey> keySelector, Func<TSource, TElement> elementSelector, IEqualityComparer<TKey> comparer)
+        {
+            Dictionary<TKey, TElement> d = new Dictionary<TKey, TElement>(source.Count, comparer);
+            foreach (TSource element in source) d.Add(keySelector(element), elementSelector(element));
+            return d;
+        }
+
 
         public static ILookup<TKey, TSource> ToLookup<TSource, TKey>(this IEnumerable<TSource> source, Func<TSource, TKey> keySelector)
         {