This PR is for the issue #3337.

Performance tests (_ToDictionary_Performance_ from _System.Linq.Tests_).
- Original - time of execution of the original code;
- Capacity est. - time of execution after capacity estimation for Array and ICollection was added;
- No IdentityFunction (ms) - time of execution after special implementation of the overload without _elementSelector_ was introduced (includes capacity estimation tweak).

__1000000 elements in source collection (100 iterations)__

Source collection | Original (ms) | Capacity est. (ms) | Boost (%) | No IdentityFunction (ms) | Boost (%)
----------------------|------------------|-------------------------|--------------|-------------------|-------
Array | 3604,5024ms | 2276,9344ms | -37% | 2088,8203ms | -42%
IEnumerable | 3598,2668ms | 3595,5406ms | 0% | 3424,7356ms | -5%
ICollection | 3591,0245ms | 2288,1238ms | -36% | 2118,0754ms | -41%


__100 elements (1000000 iterations)__

Source collection | Original (ms) | Capacity est. (ms) | Boost (%) | No IdentityFunction (ms) | Boost (%)
----------------------|------------------|------------------------|---------------|-------------------------------|--------
Array | 3289,5846ms | 2011,3236ms | -39% | 1774,5912ms | -46%
IEnumerable | 3321,1946ms | 3361,5469ms | +1% | 3104,9243ms | -6%
ICollection | 3323,8798ms | 2049,7582ms | -38% | 1808,4537ms | -46%


__4 elements (10000000 iterations)__

Source collection | Original (ms) | Capacity est. (ms) | Boost (%) | No IdentityFunction (ms) | Boost (%)
----------------------|------------------|------------------------|---------------|-------------------------------|--------
Array | 1993,3899ms | 1480,8956ms | -26% | 1336,9621ms | -33%
IEnumerable | 2322,84ms | 2404,4192ms | +3% | 2281,7706ms | -2%
ICollection | 2327,5683ms | 1776,3507ms | -24% | 1663,1914ms | -29%


__1 element (100000000 iterations)__ /absolute time of a single iteration is very small/

Source collection | Original (ms) | Capacity est. (ms) | Boost (%) | No IdentityFunction (ms) | Boost (%)
----------------------|------------------|------------------------|---------------|-------------------------------|--------
Array | 8167,0357ms | 8858,9957ms | +8% | 8287,586ms | +1%
IEnumerable | 11219,7955ms | 12047,7075ms | +7% | 11446,0032ms | +2%
ICollection | 11228,2195ms | 12036,7601ms | +7% | 11382,9351ms | +1%


__0 elements (100000000 iterations)__ /absolute time of a single iteration is very small/

Source collection | Original (ms) | Capacity est. (ms) | Boost (%) | No IdentityFunction (ms) | Boost (%)
----------------------|------------------|------------------------|---------------|-------------------------------|--------
Array | 4042,6531ms | 4419,2022ms | +9% | 4192,6036ms | +3%
IEnumerable | 6801,501ms | 7312,8057ms | +7% | 7154,1184ms | +5%
ICollection | 6782,3701ms | 7396,2199ms | +9% | 7164,335ms | +5%