@ikopylov 2015-09-22
I have updated this PR according to the review.


__1000000 elements (100 iterations)__

Source collection | Original (ms) | This PR (ms) | Boost (%)
------------------|---------------|--------------|----------
int[] | 4686,07ms | 2162,04ms | -54%
List\<int\> | 4771,15ms | 2276,67ms | -52%
HashSet\<int\> | 4530,98ms | 2620,79ms | -42%
List\<int\>.AsReadOnly() | 4510,33ms | 2615,07ms | -42%
ICollection\<int\> | 4769,64ms | 2722,54ms | -43%
int[].Where(o =\> true) | 4874,03ms | 4695,33ms | -4%
int[].Select(o =\> o) | 4876,16ms | 4687,45ms | -4%
IEnumerable\<int\> | 4771,98ms | 4599,11ms | -4%
IEnumerable\<string\> | 7198,07ms | 6997,57ms | -3%
IEnumerable\<object\> | 7234,04ms | 6973,06ms | -4%


__100 elements (1000000 iterations)__

Source collection | Original (ms) | This PR (ms) | Boost (%)
------------------|---------------|--------------|----------
int[] | 3370,17ms | 1334,99ms | -60%
List\<int\> | 3498,47ms | 1450,50ms | -59%
HashSet\<int\> | 3591,95ms | 2053,84ms | -43%
List\<int\>.AsReadOnly() | 3481,31ms | 1927,73ms | -45%
ICollection\<int\> | 3416,00ms | 1943,26ms | -43%
int[].Where(o =\> true) | 3557,66ms | 3340,35ms | -6%
int[].Select(o =\> o) | 3555,93ms | 3354,22ms | -6%
IEnumerable\<int\> | 3485,96ms | 3349,27ms | -4%
IEnumerable\<string\> | 4115,23ms | 3859,24ms | -6%
IEnumerable\<object\> | 4167,66ms | 3941,77ms | -5%


__4 elements (10000000 iterations)__

Source collection | Original (ms) | This PR (ms) | Boost (%)
------------------|---------------|--------------|----------
int[] | 2084,06ms | 1359,89ms | -35%
List\<int\> | 2230,92ms | 1275,37ms | -43%
HashSet\<int\> | 2247,16ms | 1619,94ms | -28%
List\<int\>.AsReadOnly() | 2224,59ms | 1600,45ms | -28%
ICollection\<int\> | 2400,97ms | 1775,12ms | -26%
int[].Where(o =\> true) | 2237,62ms | 2209,26ms | -1%
int[].Select(o =\> o) | 2260,38ms | 2211,32ms | -2%
IEnumerable\<int\> | 2409,73ms | 2352,72ms | -2%
IEnumerable\<string\> | 2691,22ms | 2586,51ms | -4%
IEnumerable\<object\> | 2710,48ms | 2586,72ms | -5%


__1 elements (100000000 iterations)__

Source collection | Original (ms) | This PR (ms) | Boost (%)
------------------|---------------|--------------|----------
int[] | 8602,86ms | 9514,90ms | 11%
List\<int\> | 9333,54ms | 8207,48ms | -12%
HashSet\<int\> | 9151,18ms | 9621,01ms | 5%
List\<int\>.AsReadOnly() | 9382,98ms | 10221,59ms | 9%
ICollection\<int\> | 11649,64ms | 12130,48ms | 4%
int[].Where(o =\> true) | 9307,89ms | 9612,06ms | 3%
int[].Select(o =\> o) | 9646,70ms | 10083,64ms | 5%
IEnumerable\<int\> | 11538,08ms | 11642,09ms | 1%
IEnumerable\<string\> | 12672,73ms | 12142,82ms | -4%
IEnumerable\<object\> | 12697,94ms | 12208,75ms | -4%


__0 elements (100000000 iterations)__

Source collection | Original (ms) | This PR (ms) | Boost (%)
------------------|---------------|--------------|----------
int[] | 4072,35ms | 5137,68ms | 26%
List\<int\> | 4646,03ms | 3079,56ms | -34%
HashSet\<int\> | 4265,61ms | 3038,15ms | -29%
List\<int\>.AsReadOnly() | 4751,50ms | 3160,69ms | -33%
ICollection\<int\> | 6914,09ms | 2987,52ms | -57%
int[].Where(o =\> true) | 4917,60ms | 5341,12ms | 9%
int[].Select(o =\> o) | 5319,72ms | 5853,02ms | 10%
IEnumerable\<int\> | 6888,49ms | 7068,80ms | 3%
IEnumerable\<string\> | 7475,56ms | 7476,68ms | 0%
IEnumerable\<object\> | 7529,45ms | 7566,17ms | 0%