@ikopylov 2015-09-22
The situation with array is very interesting. Tests show performance drop for small arrays (1 element ~11%, 0 element ~26%). The cast from T[] to ICollection\<T\> is slow.

What is more interesting, if I place the cast to array before cast to ```ICollection<T>```, this degradation will disappear while other numbers remain almost the same.

```C#
TSource[] array = source as TSource[];
if (array != null)
    return ToDictionary(array, keySelector, comparer);

int capacity = 0;
ICollection<TSource> collection = source as ICollection<TSource>;
if (collection != null)
{
    capacity = collection.Count;
    if (capacity == 0)
        return new Dictionary<TKey, TSource>(comparer);

    List<TSource> list = collection as List<TSource>;
    if (list != null)
        return ToDictionary(list, keySelector, comparer);
}

Dictionary<TKey, TSource> d = new Dictionary<TKey, TSource>(capacity, comparer);
foreach (TSource element in source) d.Add(keySelector(element), element);
return d;
```

Performance tests:

1 elements (100000000 iterations)

Source collection | Original (ms) | This PR (ms) | Boost (%)
------------------|---------------|--------------|----------
int[] | 8097,72ms | 6413,20ms | -21%
List\<int\> | 9532,58ms | 8010,96ms | -16%
HashSet\<int\> | 9144,07ms | 9205,86ms | 1%
List\<int\>.AsReadOnly() | 9265,29ms | 10003,97ms | 8%
ICollection\<int\> | 11499,05ms | 12031,62ms | 5%
int[].Where(o =\> true) | 9115,60ms | 9565,83ms | 5%
int[].Select(o =\> o) | 9542,34ms | 9847,30ms | 3%
IEnumerable\<int\> | 11476,08ms | 11849,48ms | 3%
IEnumerable\<string\> | 12582,02ms | 13093,70ms | 4%
IEnumerable\<object\> | 12562,69ms | 13097,15ms | 4%


0 elements (100000000 iterations)

Source collection | Original (ms) | This PR (ms) | Boost (%)
------------------|---------------|--------------|----------
int[] | 4072,29ms | 3046,66ms | -25%
List\<int\> | 4701,58ms | 3317,23ms | -29%
HashSet\<int\> | 4429,96ms | 3264,19ms | -26%
List\<int\>.AsReadOnly() | 4747,34ms | 3451,75ms | -27%
ICollection\<int\> | 6872,32ms | 3285,33ms | -52%
int[].Where(o =\> true) | 4846,33ms | 5390,64ms | 11%
int[].Select(o =\> o) | 5051,80ms | 5633,37ms | 12%
IEnumerable\<int\> | 6849,66ms | 7145,96ms | 4%
IEnumerable\<string\> | 7385,51ms | 7493,31ms | 1%
IEnumerable\<object\> | 7330,84ms | 7482,88ms | 2%