@stephentoub 2015-09-23
Other than the minor nits, LGTM.

cc: @VSadov